refactor: rename streamlit app package
