+++ streamlit_app/__init__.py
@@ -1,0 +1,3 @@
+"""OpenSells Streamlit app package."""
+
+# This file makes ``streamlit_app`` a Python package.
