+++ streamlit_app/app.py
@@ -3,9 +3,9 @@
 import os
 import streamlit as st
 
-from streamlit.auth_utils import ensure_token_and_user, logout_button
-from streamlit.cookies_utils import init_cookie_manager_mount
-from streamlit.utils import http_client
+from streamlit_app.auth_utils import ensure_token_and_user, logout_button
+from streamlit_app.cookies_utils import init_cookie_manager_mount
+from streamlit_app.utils import http_client
 
 init_cookie_manager_mount()
