+++ streamlit_app/pages/4_Tareas.py
@@ -6,11 +6,11 @@ import time
 from datetime import date
 from dotenv import load_dotenv
 
-from streamlit.cache_utils import cached_get, cached_post, limpiar_cache
-from streamlit.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
-from streamlit.auth_utils import ensure_token_and_user, logout_button
-from streamlit.cookies_utils import init_cookie_manager_mount
-from streamlit.utils import http_client
+from streamlit_app.cache_utils import cached_get, cached_post, limpiar_cache
+from streamlit_app.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
+from streamlit_app.auth_utils import ensure_token_and_user, logout_button
+from streamlit_app.cookies_utils import init_cookie_manager_mount
+from streamlit_app.utils import http_client
 
 init_cookie_manager_mount()
 # ────────────────── Config ──────────────────────────
