+++ streamlit_app/pages/8_Mi_Cuenta.py
@@ -8,11 +8,11 @@ import io
 from dotenv import load_dotenv
 from json import JSONDecodeError
 
-from streamlit.cache_utils import cached_get, cached_post, limpiar_cache
-from streamlit.auth_utils import ensure_token_and_user, logout_button
-from streamlit.utils import http_client
-from streamlit.plan_utils import subscription_cta, force_redirect
-from streamlit.cookies_utils import init_cookie_manager_mount
+from streamlit_app.cache_utils import cached_get, cached_post, limpiar_cache
+from streamlit_app.auth_utils import ensure_token_and_user, logout_button
+from streamlit_app.utils import http_client
+from streamlit_app.plan_utils import subscription_cta, force_redirect
+from streamlit_app.cookies_utils import init_cookie_manager_mount
 
 init_cookie_manager_mount()
